trigger abort when we cannot resync
