+++ BaragonAgentService/src/main/java/com/hubspot/baragon/agent/managed/BootstrapManaged.java
@@ -59,32 +59,24 @@ public class BootstrapManaged implements Managed {
 
   private final BaragonAgentConfiguration configuration;
   private final BaragonLoadBalancerDatastore loadBalancerDatastore;
-  private final BaragonAuthDatastore authDatastore;
   private final LeaderLatch leaderLatch;
   private final BaragonKnownAgentsDatastore knownAgentsDatastore;
   private final BaragonAgentMetadata baragonAgentMetadata;
   private final ScheduledExecutorService executorService;
   private final AgentHeartbeatWorker agentHeartbeatWorker;
-  private final BaragonWorkerDatastore workerDatastore;
-  private final HttpClient httpClient;
-  private final ResyncListener resyncListener;
-
-  private static final String SERVICE_CHECKIN_URL_FORMAT = "%s/checkin/%s/%s";
+  private final LifecycleHelper lifecycleHelper;
 
   private ScheduledFuture<?> requestWorkerFuture = null;
 
   @Inject
   public BootstrapManaged(BaragonKnownAgentsDatastore knownAgentsDatastore,
                           BaragonLoadBalancerDatastore loadBalancerDatastore,
-                          BaragonWorkerDatastore workerDatastore,
                           BaragonAgentConfiguration configuration,
-                          BaragonAuthDatastore authDatastore,
                           AgentHeartbeatWorker agentHeartbeatWorker,
                           BaragonAgentMetadata baragonAgentMetadata,
-                          ResyncListener resyncListener,
+                          LifecycleHelper lifecycleHelper,
                           @Named(BaragonAgentServiceModule.AGENT_SCHEDULED_EXECUTOR) ScheduledExecutorService executorService,
-                          @Named(BaragonAgentServiceModule.AGENT_LEADER_LATCH) LeaderLatch leaderLatch,
-                          @Named(BaragonAgentServiceModule.BARAGON_AGENT_HTTP_CLIENT) HttpClient httpClient) {
+                          @Named(BaragonAgentServiceModule.AGENT_LEADER_LATCH) LeaderLatch leaderLatch) {
     this.configuration = configuration;
     this.leaderLatch = leaderLatch;
     this.knownAgentsDatastore = knownAgentsDatastore;
@@ -92,10 +84,7 @@ public class BootstrapManaged implements Managed {
     this.baragonAgentMetadata = baragonAgentMetadata;
     this.executorService = executorService;
     this.agentHeartbeatWorker = agentHeartbeatWorker;
-    this.workerDatastore = workerDatastore;
-    this.authDatastore = authDatastore;
-    this.httpClient = httpClient;
-    this.resyncListener = resyncListener;
+    this.lifecycleHelper = lifecycleHelper;
   }
 
 
@@ -103,14 +92,14 @@ public class BootstrapManaged implements Managed {
   @Override
   public void start() throws Exception {
     LOG.info("Applying current configs...");
-    resyncListener.applyCurrentConfigs();
+    lifecycleHelper.applyCurrentConfigs();
 
     LOG.info("Starting leader latch...");
     leaderLatch.start();
 
     if (configuration.isRegisterOnStartup()) {
       LOG.info("Notifying BaragonService...");
-      notifyService("startup");
+      lifecycleHelper.notifyService("startup");
     }
 
     LOG.info("Updating BaragonGroup information...");
@@ -122,83 +111,11 @@ public class BootstrapManaged implements Managed {
     LOG.info("Starting agent heartbeat...");
     requestWorkerFuture = executorService.scheduleAtFixedRate(agentHeartbeatWorker, 0, configuration.getHeartbeatIntervalSeconds(), TimeUnit.SECONDS);
 
-    if (configuration.getStateFile().isPresent()) {
-      LOG.info("Writing state file...");
-      writeStateFile();
-    }
+    lifecycleHelper.writeStateFileIfConfigured();
   }
 
   @Override
   public void stop() throws Exception {
-    leaderLatch.close();
-    executorService.shutdown();
-    if (configuration.isDeregisterOnGracefulShutdown()) {
-      LOG.info("Notifying BaragonService of shutdown...");
-      notifyServiceWithRetry("shutdown");
-    }
-    if (configuration.getStateFile().isPresent()) {
-      removeStateFile();
-    }
-  }
-
-  private void notifyServiceWithRetry(final String action) {
-    Callable<Void> callable = new Callable<Void>() {
-      public Void call() throws Exception {
-        notifyService(action);
-        return null;
-      }
-    };
-
-    Retryer<Void> retryer = RetryerBuilder.<Void>newBuilder()
-      .retryIfException()
-      .withStopStrategy(StopStrategies.stopAfterAttempt(configuration.getMaxNotifyServiceAttempts()))
-      .withWaitStrategy(WaitStrategies.exponentialWait(1, TimeUnit.SECONDS))
-      .build();
-
-    try {
-      retryer.call(callable);
-    } catch (Exception e) {
-      if (action.equals("startup") && !configuration.isExitOnStartupError()) {
-        LOG.error("Could not notify service of startup", e);
-      } else {
-        throw Throwables.propagate(e);
-      }
-    }
-  }
-
-  private void notifyService(String action) throws AgentStartupException {
-    Collection<String> baseUris = workerDatastore.getBaseUris();
-    if (!baseUris.isEmpty()) {
-      HttpRequest.Builder requestBuilder = HttpRequest.newBuilder()
-        .setUrl(String.format(SERVICE_CHECKIN_URL_FORMAT, baseUris.iterator().next(), configuration.getLoadBalancerConfiguration().getName(), action))
-        .setMethod(HttpRequest.Method.POST)
-        .setBody(baragonAgentMetadata);
-
-      Map<String, BaragonAuthKey> authKeys = authDatastore.getAuthKeyMap();
-      if (!authKeys.isEmpty()) {
-        requestBuilder.setQueryParam("authkey").to(authKeys.entrySet().iterator().next().getValue().getValue());
-      }
-
-      HttpRequest request = requestBuilder.build();
-      HttpResponse response = httpClient.execute(request);
-      LOG.info(String.format("Got %s response from BaragonService", response.getStatusCode()));
-      if (response.isError()) {
-        throw new AgentStartupException(String.format("Bad response received from BaragonService %s", response.getAsString()));
-      }
-    }
-  }
-
-  private void writeStateFile() throws IOException {
-    Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(configuration.getStateFile().get()), "UTF-8"));
-    try {
-      writer.write("RUNNING");
-    } finally {
-      writer.close();
-    }
-  }
-
-  private boolean removeStateFile() {
-    File stateFile = new File(configuration.getStateFile().get());
-    return (!stateFile.exists() || stateFile.delete());
+    lifecycleHelper.shutdown();
   }
 }
